requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
elastic-transport==8.1.2
elasticsearch==8.3.3
emoji==2.0.0
pycodestyle==2.9.1
pydantic==1.9.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
telegram==0.0.1
toml==0.10.2
tornado==6.2
typing-extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
